"simple changes in view file"
